moved game into seperate file
